refactor(ui): split integrations overlay into focused modules
